# docs/swagger.yaml
definitions:
...
  models.ResponseData:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  models.TestingAnswerItem:
    properties:
      answer_list:
        items: {}
        type: array
      part_name:
        type: string
    type: object
...
  models.TestingRecordsItem:
    properties:
      answers:
        items:
          $ref: '#/definitions/models.TestingAnswerItem'
        type: array
      id:
        type: integer
      name:
        type: string
      rest_seconds:
        items:
          type: integer
        type: array
      score:
        items:
          type: integer
        type: array
      status:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  models.TestingRecordsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.TestingRecordsItem'
        type: array
      total:
        type: integer
    type: object
...
  models.WritingRecordsItem:
    properties:
      answers:
        items: {}
        type: array
      id:
        type: integer
      name:
        type: string
      rest_seconds:
        type: integer
      status:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  models.WritingRecordsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.WritingRecordsItem'
        type: array
      total:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /record/testing/add:
    post:
      consumes:
      - application/json
      description: 新增测试做题记录
      parameters:
      - description: 测试做题记录内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.TestingRecordsItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增测试做题记录
      tags:
      - Testing
  /record/testing/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除测试做题记录
      parameters:
      - description: 测试做题记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除测试做题记录
      tags:
      - Testing
  /record/testing/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据id获取测试做题记录详情
      parameters:
      - description: 测试做题记录id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.TestingRecordsItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取测试做题记录详情
      tags:
      - Testing
  /record/testing/list:
    get:
      consumes:
      - application/json
      description: 根据条件获取测试记录列表，并返回分页结果
      parameters:
      - description: 名称
        in: query
        name: name
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 试题类型
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageLimit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.TestingRecordsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取测试做题记录列表
      tags:
      - Testing
  /record/testing/update:
    put:
      consumes:
      - application/json
      description: 更新测试做题记录
      parameters:
      - description: 测试做题记录内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.TestingRecordsItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 更新测试做题记录
      tags:
      - Testing
  /record/writing/add:
    post:
      consumes:
      - application/json
      description: 新增写作做题记录
      parameters:
      - description: 写作做题记录内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.WritingRecordsItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增写作做题记录
      tags:
      - Writing
  /record/writing/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除写作做题记录
      parameters:
      - description: 写作做题记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除写作做题记录
      tags:
      - Writing
  /record/writing/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据id获取写作做题记录详情
      parameters:
      - description: 写作做题记录id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.WritingRecordsItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取写作做题记录详情
      tags:
      - Writing
  /record/writing/list:
    get:
      consumes:
      - application/json
      description: 根据条件获取写作记录列表，并返回分页结果
      parameters:
      - description: 名称
        in: query
        name: name
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 试题类型
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageLimit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.WritingRecordsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取写作做题记录列表
      tags:
      - Writing
  /record/writing/update:
    put:
      consumes:
      - application/json
      description: 更新写作做题记录
      parameters:
      - description: 写作做题记录内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.WritingRecordsItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 更新写作做题记录
      tags:
      - Writing
...
swagger: "2.0"